requirements.txt
django>=3.2
djangorestframework>=3.12.4
cryptography==3.4.7
mysql-connector-python==8.0.25
numpy==1.20.3
pandas==1.2.4
PyMySQL==1.0.2